+++ rotkehlchen/exchanges/coinbasepro.py
@@ -1,28 +1,15 @@
 import binascii
-import csv
 import hashlib
 import hmac
 import json
 import logging
-import os
 import time
 from base64 import b64decode, b64encode
 from collections import defaultdict
 from http import HTTPStatus
 from json.decoder import JSONDecodeError
-from tempfile import TemporaryDirectory
-from typing import (
-    TYPE_CHECKING,
-    Any,
-    Callable,
-    DefaultDict,
-    Dict,
-    List,
-    Optional,
-    Tuple,
-    Union,
-    overload,
-)
+from typing import TYPE_CHECKING, Any, DefaultDict, Dict, Iterator, List, Optional, Tuple, Union
+from urllib.parse import urlencode
 
 import gevent
 import requests
@@ -31,6 +18,7 @@ from typing_extensions import Literal
 from rotkehlchen.accounting.structures import Balance
 from rotkehlchen.assets.asset import Asset
 from rotkehlchen.assets.converters import asset_from_coinbase
+from rotkehlchen.constants.assets import A_ETH
 from rotkehlchen.constants.misc import ZERO
 from rotkehlchen.constants.timing import QUERY_RETRY_TIMES
 from rotkehlchen.errors import (
@@ -52,11 +40,20 @@ from rotkehlchen.serialization.deserialize import (
     deserialize_price,
     deserialize_timestamp_from_date,
     deserialize_trade_type,
+    pair_get_assets,
 )
-from rotkehlchen.typing import ApiKey, ApiSecret, Fee, Location, Timestamp, TradePair
+from rotkehlchen.typing import (
+    ApiKey,
+    ApiSecret,
+    AssetMovementCategory,
+    AssetType,
+    Fee,
+    Location,
+    Timestamp,
+    TradePair,
+)
 from rotkehlchen.user_messages import MessagesAggregator
 from rotkehlchen.utils.interfaces import cache_response_timewise, protect_with_lock
-from rotkehlchen.utils.misc import timestamp_to_iso8601, ts_now
 from rotkehlchen.utils.serialization import jsonloads_dict, jsonloads_list
 
 if TYPE_CHECKING:
@@ -66,8 +63,7 @@ logger = logging.getLogger(__name__)
 log = RotkehlchenLogsAdapter(logger)
 
 
-SECS_TO_WAIT_FOR_REPORT = 300
-MINS_TO_WAIT_FOR_REPORT = SECS_TO_WAIT_FOR_REPORT / 60
+COINBASEPRO_PAGINATION_LIMIT = 100  # default + max limit
 
 
 def coinbasepro_to_worldpair(product: str) -> TradePair:
@@ -90,6 +86,20 @@ class CoinbaseProPermissionError(Exception):
     pass
 
 
+def coinbasepro_deserialize_timestamp(entry: Dict[str, Any], key: str) -> Timestamp:
+    """Deserialize a timestamp from coinbasepro
+
+    In case of an error raises:
+    - KeyError
+    - DeserializationError
+    """
+    raw_time = entry[key]
+    if raw_time.endswith('+00'):  # proper iso8601 needs + 00:00 for timezone
+        raw_time = raw_time.replace('+00', '+00:00')
+    timestamp = deserialize_timestamp_from_date(raw_time, 'iso8601', 'coinbasepro')
+    return timestamp
+
+
 class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
 
     def __init__(
@@ -103,6 +113,7 @@ class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
         super().__init__('coinbasepro', api_key, secret, database)
         self.base_uri = 'https://api.pro.coinbase.com'
         self.msg_aggregator = msg_aggregator
+        self.account_to_currency: Optional[Dict[str, Asset]] = None
 
         self.session.headers.update({
             'Content-Type': 'Application/JSON',
@@ -138,42 +149,18 @@ class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
 
         return True, ''
 
-    @overload
-    def _api_query(  # pylint: disable=no-self-use
-            self,
-            endpoint: Literal['accounts', 'products'],
-            request_method: Literal['GET'] = 'GET',
-            options: Optional[Dict[str, Any]] = None,
-    ) -> List[Any]:
-        ...
-
-    @overload
-    def _api_query(  # pylint: disable=no-self-use
-            self,
-            endpoint: Literal['reports'],
-            request_method: Literal['GET', 'POST'] = 'GET',
-            options: Optional[Dict[str, Any]] = None,
-    ) -> Dict[str, Any]:
-        ...
-
-    @overload
-    def _api_query(  # pylint: disable=no-self-use
-            self,
-            endpoint: str,
-            request_method: Literal['GET', 'POST'] = 'GET',
-            options: Optional[Dict[str, Any]] = None,
-    ) -> Union[List[Any], Dict[str, Any]]:
-        ...
-
     def _api_query(
             self,
             endpoint: str,
             request_method: Literal['GET', 'POST'] = 'GET',
             options: Optional[Dict[str, Any]] = None,
-    ) -> Union[List[Any], Dict[str, Any]]:
+            query_options: Optional[Dict[str, Any]] = None,
+    ) -> Tuple[List[Any], Optional[str]]:
         """Performs a coinbase PRO API Query for endpoint
 
         You can optionally provide extra arguments to the endpoint via the options argument.
+
+        Returns a tuple of the result and optional pagination cursor.
 
         Raises RemoteError if something went wrong with connecting or reading from the exchange
         Raises CoinbaseProPermissionError if the API Key does not have sufficient
@@ -187,6 +174,10 @@ class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
         else:
             stringified_options = ''
             options = {}
+
+        if query_options:
+            request_url += '?' + urlencode(query_options)
+
         message = timestamp + request_method + request_url + stringified_options
         log.debug(
             'Coinbase Pro API query',
@@ -194,6 +185,7 @@ class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
             request_url=request_url,
             options=options,
         )
+
         if 'products' not in endpoint:
             try:
                 signature = hmac.new(
@@ -253,27 +245,52 @@ class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
                 f'status code: {response.status_code} and text: {response.text}',
             )
 
-        loading_function: Union[Callable[[str], Dict[str, Any]], Callable[[str], List[Any]]]
-        if any(x in endpoint for x in ('accounts', 'products')):
-            loading_function = jsonloads_list
-        else:
-            loading_function = jsonloads_dict
-
         try:
-            json_ret = loading_function(response.text)
+            json_ret = jsonloads_list(response.text)
         except JSONDecodeError as e:
             raise RemoteError(
                 f'Coinbase Pro {request_method} query at {full_url} '
                 f'returned invalid JSON response: {response.text}',
             ) from e
 
-        return json_ret
+        return json_ret, response.headers.get('cb-after', None)
+
+    def create_or_return_account_to_currency_map(self) -> Dict[str, Asset]:
+        if self.account_to_currency is not None:
+            return self.account_to_currency
+
+        accounts, _ = self._api_query('accounts')
+        self.account_to_currency = {}
+        for account in accounts:
+            try:
+                asset = asset_from_coinbase(account['currency'])
+                self.account_to_currency[account['id']] = asset
+            except UnsupportedAsset as e:
+                self.msg_aggregator.add_warning(
+                    f'Found coinbase pro account with unsupported asset '
+                    f'{e.asset_name}. Ignoring it.',
+                )
+                continue
+            except UnknownAsset as e:
+                self.msg_aggregator.add_warning(
+                    f'Found coinbase pro account result with unknown asset '
+                    f'{e.asset_name}. Ignoring it.',
+                )
+                continue
+            except KeyError as e:
+                self.msg_aggregator.add_warning(
+                    f'Found coinbase pro account entry with missing {str(e)} field. '
+                    f'Ignoring it',
+                )
+                continue
+
+        return self.account_to_currency
 
     @protect_with_lock()
     @cache_response_timewise()
     def query_balances(self) -> ExchangeQueryBalances:
         try:
-            accounts = self._api_query('accounts')
+            accounts, _ = self._api_query('accounts')
         except (CoinbaseProPermissionError, RemoteError) as e:
             msg = f'Coinbase Pro API request failed. {str(e)}'
             log.error(msg)
@@ -331,261 +348,111 @@ class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
 
         return dict(assets_balance), ''
 
-    def _get_products_ids(self) -> List[str]:
-        """Gets a list of all product ids (markets) offered by coinbase PRO
-
-        - Raises the same exceptions as _api_query()
-        - Can raise KeyError if the API does not return the expected response format.
-        """
-        products = self._api_query('products', request_method='GET')
-        return [product['id'] for product in products]
-
-    def _get_account_ids(self) -> List[str]:
-        """Gets a list of all account ids
-
-        - Raises the same exceptions as _api_query()
-        - Can raise KeyError if the API does not return the expected response format.
-        """
-        accounts = self._api_query('accounts')
-        return [account['id'] for account in accounts]
-
-    def _read_asset_movements(self, filepath: str) -> List[AssetMovement]:
-        """Reads a csv account type report and extracts the AssetMovements"""
-        with open(filepath, newline='') as csvfile:
-            reader = csv.DictReader(csvfile)
-            movements = []
-            for row in reader:
-                try:
-                    if row['type'] in ('withdrawal', 'deposit'):
-                        timestamp = deserialize_timestamp_from_date(
-                            row['time'],
-                            'iso8601',
-                            'coinbasepro',
-                        )
-                        asset = Asset(row['amount/balance unit'])
-                        movements.append(AssetMovement(
-                            location=Location.COINBASEPRO,
-                            category=deserialize_asset_movement_category(row['type']),
-                            address=None,  # can't get it from csv data
-                            transaction_id=None,  # can't get it from csv data
-                            timestamp=timestamp,
-                            asset=asset,
-                            amount=deserialize_asset_amount_force_positive(row['amount']),
-                            fee_asset=asset,
-                            # I don't see any fee in deposit withdrawals in coinbasepro
-                            fee=Fee(ZERO),
-                            link=str(row['transfer id']),
-                        ))
-                except UnknownAsset as e:
-                    self.msg_aggregator.add_warning(
-                        f'Found unknown Coinbasepro asset {e.asset_name}. '
-                        f'Ignoring its deposit/withdrawal.',
-                    )
-                    continue
-                except (DeserializationError, KeyError) as e:
-                    msg = str(e)
-                    if isinstance(e, KeyError):
-                        msg = f'Missing key entry for {msg}.'
-                    self.msg_aggregator.add_error(
-                        'Failed to deserialize a Coinbasepro deposit/withdrawal. '
-                        'Check logs for details. Ignoring it.',
-                    )
-                    log.error(
-                        'Error processing a coinbasepro  deposit/withdrawal.',
-                        raw_asset_movement=row,
-                        error=msg,
-                    )
-                    continue
-
-        return movements
-
-    def _read_trades(self, filepath: str) -> List[Trade]:
-        """Reads a csv fill type report and extracts the Trades"""
-        with open(filepath, newline='') as csvfile:
-            reader = csv.DictReader(csvfile)
-            trades = []
-            for row in reader:
-                try:
-                    timestamp = deserialize_timestamp_from_date(
-                        row['created at'],
-                        'iso8601',
-                        'coinbasepro',
-                    )
-                    trades.append(Trade(
-                        timestamp=timestamp,
-                        location=Location.COINBASEPRO,
-                        pair=coinbasepro_to_worldpair(row['product']),
-                        trade_type=deserialize_trade_type(row['side']),
-                        amount=deserialize_asset_amount(row['size']),
-                        rate=deserialize_price(row['price']),
-                        fee=deserialize_fee(row['fee']),
-                        fee_currency=Asset(row['price/fee/total unit']),
-                        link=row['trade id'],
-                        notes='',
-                    ))
-                except UnprocessableTradePair as e:
-                    self.msg_aggregator.add_warning(
-                        f'Found unprocessable Coinbasepro pair {e.pair}. Ignoring the trade.',
-                    )
-                    continue
-                except UnknownAsset as e:
-                    self.msg_aggregator.add_warning(
-                        f'Found unknown Coinbasepro asset {e.asset_name}. '
-                        f'Ignoring the trade.',
-                    )
-                    continue
-                except (DeserializationError, KeyError) as e:
-                    msg = str(e)
-                    if isinstance(e, KeyError):
-                        msg = f'Missing key entry for {msg}.'
-                    self.msg_aggregator.add_error(
-                        'Failed to deserialize a coinbasepro trade. '
-                        'Check logs for details. Ignoring it.',
-                    )
-                    log.error(
-                        'Error processing a coinbasepro trade.',
-                        raw_trade=row,
-                        error=msg,
-                    )
-                    continue
-
-        return trades
-
-    def _generate_reports(
+    def _paginated_query(
             self,
-            start_ts: Timestamp,
-            end_ts: Timestamp,
-            report_type: Literal['fills', 'account'],
-            tempdir: str,
-    ) -> List[str]:
-        """
-        Generates all the reports to get historical data from coinbase.
-
-        https://docs.pro.coinbase.com/#reports
-        There are 2 type of reports.
-        1. Fill reports which are per product id (market)
-        2. Account reports which are per account id
-
-        The fill reports have the following data format:
-        portfolio,trade id,product,side,created at,size,size unit,price,fee,
-        total,price/fee/total unit
-
-        The account reports have the following data format:
-        portfolio,type,time,amount,balance,amount/balance unit,transfer id,trade id,order id
-
-        Returns a list of filepaths where the reports were written.
-
-        - Raises the same exceptions as _api_query()
-        - Can raise KeyError if the API does not return the expected response format.
-        """
-        start_date = timestamp_to_iso8601(start_ts)
-        end_date = timestamp_to_iso8601(end_ts)
-
-        if report_type == 'fills':
-            account_or_product_ids = self._get_products_ids()
-            identifier_key = 'product_id'
-        else:
-            account_or_product_ids = self._get_account_ids()
-            identifier_key = 'account_id'
-
-        report_ids = []
-        options = {
-            'type': report_type,
-            'start_date': start_date,
-            'end_date': end_date,
-            'format': 'csv',
-            # The only way to disable emailing the report link is to give an invalid link
-            'email': 'some@invalidemail.com',
-        }
-        for identifier in account_or_product_ids:
-            options[identifier_key] = identifier
-            post_result = self._api_query('reports', request_method='POST', options=options)
-            report_ids.append(post_result['id'])
-
-        # At this point all reports must have been queued for creation at the server
-        # Now wait until they are ready and pull them one by one
-        report_paths = []
-        last_change_ts = ts_now()
+            endpoint: str,
+            query_options: Dict[str, Any] = None,
+            limit: int = COINBASEPRO_PAGINATION_LIMIT,
+    ) -> Iterator[List[Dict[str, Any]]]:
+        if query_options is None:
+            query_options = {}
+        query_options['limit'] = limit
         while True:
-            finished_ids_indices = []
-            for idx, report_id in enumerate(report_ids):
-                get_result = self._api_query(f'reports/{report_id}', request_method='GET')
-                # Have to add assert here for mypy since the endpoint string is
-                # a variable string and can't be overloaded and type checked
-                assert isinstance(get_result, dict)
-                if get_result['status'] != 'ready':
-                    continue
-                # a report is ready here so let's reset the timer
-                last_change_ts = ts_now()
-                file_url = get_result['file_url']
-                response = requests.get(file_url)
-                length = len(response.content)
-                # empty fill reports have length of 95, empty account reports 85
-                # So we assume a report of more than 100 chars has data.
-                if length > 100:
-                    log.debug(f'Got a populated report for id: {report_id}. Writing it to disk')
-                    filepath = os.path.join(tempdir, f'report_{report_id}.csv')
-                    with open(filepath, 'wb') as f:
-                        f.write(response.content)
-                    report_paths.append(filepath)
-                else:
-                    log.debug(f'Got report for id: {report_id} with length {length}. Skipping it')
-
-                finished_ids_indices.append(idx)
-
-            if ts_now() - last_change_ts > SECS_TO_WAIT_FOR_REPORT:
-                raise RemoteError(
-                    f'There has been no response from CoinbasePro reports for over '
-                    f' {MINS_TO_WAIT_FOR_REPORT} minutes. Bailing out.',
-                )
-
-            # Delete the report ids that have been downloaded. Note: reverse order
-            # so that we don't mess up the indices
-            for idx in reversed(finished_ids_indices):
-                del report_ids[idx]
-
-            # When there is no more ids to query break out of the loop
-            if len(report_ids) == 0:
+            result, after_cursor = self._api_query(endpoint=endpoint, query_options=query_options)
+            yield result
+            if after_cursor is None or len(result) < limit:
                 break
 
-        return report_paths
+            query_options['after'] = after_cursor
 
     def query_online_deposits_withdrawals(
             self,
             start_ts: Timestamp,
             end_ts: Timestamp,
     ) -> List[AssetMovement]:
-        """Queries coinbase pro for asset movements
-
-        1. Generates relevant reports and writes them to a temporary directory
-        2. Reads all files from that directory and extracts deposits/withdrawals
-        3. Temporary directory is removed
-        """
+        """Queries coinbase pro for asset movements"""
         log.debug('Query coinbasepro asset movements', start_ts=start_ts, end_ts=end_ts)
         movements = []
-        with TemporaryDirectory() as tempdir:
-            try:
-                filepaths = self._generate_reports(
-                    start_ts=start_ts,
-                    end_ts=end_ts,
-                    report_type='account',
-                    tempdir=tempdir,
-                )
-            except CoinbaseProPermissionError as e:
-                self.msg_aggregator.add_error(
-                    f'Got permission error while querying Coinbasepro for '
-                    f'asset movements: {str(e)}',
-                )
-                return []
-            except RemoteError as e:
-                self.msg_aggregator.add_error(
-                    f'Got remote error while querying Coinbasepro for '
-                    f'asset movements: {str(e)}',
-                )
-                return []
+        raw_movements = []
+        for batch in self._paginated_query(
+            endpoint='transfers',
+            query_options={'type': 'withdraw'},
+        ):
+            raw_movements.extend(batch)
+        for batch in self._paginated_query(
+            endpoint='transfers',
+            query_options={'type': 'deposit'},
+        ):
+            raw_movements.extend(batch)
 
-            for filepath in filepaths:
-                movements.extend(self._read_asset_movements(filepath))
+        account_to_currency = self.create_or_return_account_to_currency_map()
+        for entry in raw_movements:
+            try:
+                timestamp = coinbasepro_deserialize_timestamp(entry, 'completed_at')
+                if timestamp < start_ts or timestamp > end_ts:
+                    continue
+
+                category = deserialize_asset_movement_category(entry['type'])
+                asset = account_to_currency.get(entry['account_id'], None)
+                if asset is None:
+                    log.warning(
+                        f'Skipping coinbase pro asset_movement {entry} due to inability to '
+                        f'match account id to an asset',
+                    )
+                    continue
+
+                address = None
+                transaction_id = None
+                fee = Fee(ZERO)
+                if category == AssetMovementCategory.DEPOSIT:
+                    try:
+                        address = entry['details']['crypto_address']
+                        transaction_id = entry['details']['crypto_transaction_hash']
+                    except KeyError:
+                        pass
+                else:  # withdrawal
+                    try:
+                        address = entry['details']['sent_to_address']
+                        transaction_id = entry['details']['crypto_transaction_hash']
+                        fee = deserialize_fee(entry['details']['fee'])
+                    except KeyError:
+                        pass
+
+                if transaction_id and (asset == A_ETH or asset.asset_type == AssetType.ETHEREUM_TOKEN):  # noqa: E501
+                    transaction_id = '0x' + transaction_id
+
+                movements.append(AssetMovement(
+                    location=Location.COINBASEPRO,
+                    category=category,
+                    address=address,
+                    transaction_id=transaction_id,
+                    timestamp=timestamp,
+                    asset=asset,
+                    amount=deserialize_asset_amount_force_positive(entry['amount']),
+                    fee_asset=asset,
+                    fee=fee,
+                    link=str(entry['id']),
+                ))
+            except UnknownAsset as e:
+                self.msg_aggregator.add_warning(
+                    f'Found unknown Coinbasepro asset {e.asset_name}. '
+                    f'Ignoring its deposit/withdrawal.',
+                )
+                continue
+            except (DeserializationError, KeyError) as e:
+                msg = str(e)
+                if isinstance(e, KeyError):
+                    msg = f'Missing key entry for {msg}.'
+                self.msg_aggregator.add_error(
+                    'Failed to deserialize a Coinbasepro deposit/withdrawal. '
+                    'Check logs for details. Ignoring it.',
+                )
+                log.error(
+                    'Error processing a coinbasepro  deposit/withdrawal.',
+                    raw_asset_movement=entry,
+                    error=msg,
+                )
+                continue
 
         return movements
 
@@ -594,35 +461,64 @@ class Coinbasepro(ExchangeInterface):  # lgtm[py/missing-call-to-init]
             start_ts: Timestamp,
             end_ts: Timestamp,
     ) -> List[Trade]:
-        """Queries coinbase pro for trades
-
-        1. Generates relevant reports and writes them to a temporary directory
-        2. Reads all files from that directory and extracts Trades
-        3. Temporary directory is removed
-        """
+        """Queries coinbase pro for trades"""
         log.debug('Query coinbasepro trade history', start_ts=start_ts, end_ts=end_ts)
-        trades = []
-        with TemporaryDirectory() as tempdir:
-            try:
-                filepaths = self._generate_reports(
-                    start_ts=start_ts,
-                    end_ts=end_ts,
-                    report_type='fills',
-                    tempdir=tempdir,
-                )
-            except CoinbaseProPermissionError as e:
-                self.msg_aggregator.add_error(
-                    f'Got permission error while querying Coinbasepro for trades: {str(e)}',
-                )
-                return []
-            except RemoteError as e:
-                self.msg_aggregator.add_error(
-                    f'Got remote error while querying Coinbasepro for trades: {str(e)}',
-                )
-                return []
 
-            for filepath in filepaths:
-                trades.extend(self._read_trades(filepath))
+        trades = []
+        raw_trades = []
+        for batch in self._paginated_query(
+            endpoint='orders',
+            query_options={'status': 'done'},
+        ):
+            raw_trades.extend(batch)
+
+        for entry in raw_trades:
+            timestamp = coinbasepro_deserialize_timestamp(entry, 'created_at')
+            if timestamp < start_ts or timestamp > end_ts:
+                continue
+
+            try:
+                pair = coinbasepro_to_worldpair(entry['product_id'])
+                # Fee currency seems to always be quote asset
+                # https://github.com/ccxt/ccxt/blob/ddf3a15cbff01541f0b37c35891aa143bb7f9d7b/python/ccxt/coinbasepro.py#L724  # noqa: E501
+                _, quote_asset = pair_get_assets(pair)
+                trades.append(Trade(
+                    timestamp=timestamp,
+                    location=Location.COINBASEPRO,
+                    pair=coinbasepro_to_worldpair(entry['product_id']),
+                    trade_type=deserialize_trade_type(entry['side']),
+                    amount=deserialize_asset_amount(entry['size']),
+                    rate=deserialize_price(entry['price']),
+                    fee=deserialize_fee(entry['fill_fees']),
+                    fee_currency=quote_asset,
+                    link=entry['id'],
+                    notes='',
+                ))
+            except UnprocessableTradePair as e:
+                self.msg_aggregator.add_warning(
+                    f'Found unprocessable Coinbasepro pair {e.pair}. Ignoring the trade.',
+                )
+                continue
+            except UnknownAsset as e:
+                self.msg_aggregator.add_warning(
+                    f'Found unknown Coinbasepro asset {e.asset_name}. '
+                    f'Ignoring the trade.',
+                )
+                continue
+            except (DeserializationError, KeyError) as e:
+                msg = str(e)
+                if isinstance(e, KeyError):
+                    msg = f'Missing key entry for {msg}.'
+                self.msg_aggregator.add_error(
+                    'Failed to deserialize a coinbasepro trade. '
+                    'Check logs for details. Ignoring it.',
+                )
+                log.error(
+                    'Error processing a coinbasepro trade.',
+                    raw_trade=entry,
+                    error=msg,
+                )
+                continue
 
         return trades